Redo the per-client-data in-flight monitoring. (#310)
* Redo the per-client-data in-flight monitoring.

The previous approach with using a global map of
pointers to shared_ptr does not work for one simple reason;
on shutdown, C++ does not define the order in which statics
are destroyed.  Thus it is possible (and even likely in some
cases) that on Ctrl-C, the global map would be deleted, and then
we would try to access it, leading to crashes.  I was seeing some
of these while running the rcl tests.

Instead, go back to the previous solution where we were storing
num_in_flight completely within the ClientData structure.
This requires doing a bit of gymnastics with the data types,
and reintrodues the possible UB during shutdown if the structure
was destructed but there is still queries in flight.  But it
otherwise fixes all of the tests for me locally.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>

* Fixes from review.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>

---------

Signed-off-by: Chris Lalancette <clalancette@gmail.com>