Don't shutdown contained entities. (#313)
It is tempting to think of the entities in the RMW
graph as a hierarchy, where an rmw_context_t contains
zero or more rmw_node_t, and rmw_node_t contains zero
or more publishers, subscriptions, clients, and services.

However, the reality is that the upper layers (particularly
rclcpp and rclpy) don't exactly view the entities like that.
More specifically, each entity is considered standalone,
that happens to have linkage to other entities in the graph.
For example, a publisher is considered to be a standalone entity
that happens to be linked to a particular node.

Because of this, it is not proper to shutdown all nodes within
a context when the context is shutdown.  The node should continue
to live on until it is shutdown.  And a similar argument goes
for the node shutdown; it should not shutdown the publishers,
subscriptions, clients, and services that are contained within it.

This manifested itself as a exception that was being thrown in
some of the tests in test_communication.  Because it is using
a loop with rclcpp::ok(), followed by a publisher->publish(),
the test would sometimes throw an exception if Ctrl-C was hit between
the rclcpp::ok() and the publish() call.  And that's because even
though the context has been shut down, the publisher is an
independent entity that should continue to exist until
the next rclcpp::ok().

The fix here is simple; don't shut "contained" entities down during
a context or node shutdown.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>