Fix UB in ClientData stuff. (#320)
The num_in_flight stuff was *still* wrong here.  First of
all, we forgot to increment num_in_flight when actually
kicking off a new query.  Once we did that, we had to
change the lock in NodeData to a recursive one, since the
call to delete_client_data from ClientData could be called
recursively.  And then finally we had to drop the ClientData
lock before the delete_client_data, since we are about to
delete ourselves and the unlock would have been UB.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>